+++ src/object-dom.ts
@@ -1,5 +1,4 @@
-export * from './base';
-export * from './dom';
-export * from './global';
-export * from './transformers';
-export * from './styles';
+export * from "./base";
+export * from "./dom";
+export * from "./transformers";
+export * from "./styles";
